requirements/base.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
dj-database-url==3.0.1
django==5.2.5
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-cuser==2017.3.16
django-debug-toolbar==6.0.0
django-filter==25.1
django-jazzmin==3.0.1
django-modeltranslation==0.19.16
django-query-counter==0.4.1
django-rosetta==0.10.2
django-timezone-field==7.1
django-tinymce==4.1.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.10
flower==2.0.1
humanize==4.12.3
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.4
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
polib==1.2.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
pygments==2.19.2
pyjwt==2.10.1
pytest==8.4.1
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
redis==6.4.0
requests==2.32.4
ruff==0.12.9
six==1.17.0
sqlparse==0.5.3
tabulate==0.9.0
tornado==6.5.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
